
\name{dataGUI}
\alias{dataGUI}


\title{dataGUI}
\description{loads the db GUI, default is sqlitestudio}

\usage{
dataGUI(path)
}


	

\examples{
testDB = system.file(package="dbiGUI", "testDB", "testDB.sqlite")
dataGUI(path = testDB) 
dataGUI() 


}


\author{MV 2012-Mar-01 2012-Mar-01 14:48:22}
















