
\name{MAKE}
\alias{MAKE}

\title{Make zip-s}

 

\examples{
\dontrun{ 
MAKE("utilsMPIO/utilsMPIO", install = TRUE) #copy the zip file to  "//ds/kempdba/SCIDB_MANAGEMENT/APPS/R_packages/BIN" AND installs from zip

require(utilsMPIO)
MAKE("utilsMPIO/utilsMPIO", install = FALSE) #copy the zip file to  "//ds/kempdba/SCIDB_MANAGEMENT/APPS/R_packages/BIN"
}
}







