
\name{dbPathGUI}
\alias{usernam}
\alias{sqliSave}
\alias{backup}


\title{random functions}
\description{undocumented functions, usually small wrappers; their use should be obvious from looking at name and code and the examples below.}

\examples{

paste("Hello Dr.", usernam())

sql2Rdate(Sys.time())

string2sql(sql2Rdate(Sys.time()))

}